fix(daemon): consolidate per-platform execute and file-info handlers so Windows honors envs, Linux timeout semantics, structured errors, and full process-tree kill, sets modifiedAt on file info, and returns 501 for code-run until ported
