+++ apps/daemon/pkg/toolbox/fs/get_file_info_windows.go
@@ -5,53 +5,11 @@
 
 package fs
 
-import (
-	"errors"
-	"fmt"
-	"net/http"
-	"os"
+import "os"
 
-	"github.com/gin-gonic/gin"
-)
-
-func GetFileInfo(c *gin.Context) {
-	path := c.Query("path")
-	if path == "" {
-		c.AbortWithError(http.StatusBadRequest, errors.New("path is required"))
-		return
-	}
-
-	info, err := getFileInfo(path)
-	if err != nil {
-		if os.IsNotExist(err) {
-			c.AbortWithError(http.StatusNotFound, err)
-			return
-		}
-		if os.IsPermission(err) {
-			c.AbortWithError(http.StatusForbidden, err)
-			return
-		}
-		c.AbortWithError(http.StatusBadRequest, err)
-		return
-	}
-
-	c.JSON(http.StatusOK, info)
-}
-
-func getFileInfo(path string) (FileInfo, error) {
-	info, err := os.Stat(path)
-	if err != nil {
-		return FileInfo{}, err
-	}
-
-	return FileInfo{
-		Name:        info.Name(),
-		Size:        info.Size(),
-		Mode:        info.Mode().String(),
-		ModTime:     info.ModTime().String(),
-		IsDir:       info.IsDir(),
-		Owner:       "",
-		Group:       "",
-		Permissions: fmt.Sprintf("%04o", info.Mode().Perm()),
-	}, nil
+// ownerGroup returns empty owner/group on Windows: file ownership lives in
+// the security descriptor (GetNamedSecurityInfo), not in os.FileInfo, and the
+// wire contract tolerates empty strings for these fields.
+func ownerGroup(os.FileInfo) (owner, group string) {
+	return "", ""
 }
