+++ apps/daemon/pkg/toolbox/process/execute.go
@@ -1,0 +1,107 @@
+// Copyright Daytona Platforms Inc.
+// SPDX-License-Identifier: AGPL-3.0
+
+package process
+
+import (
+	"bytes"
+	"errors"
+	"fmt"
+	"log/slog"
+	"net/http"
+	"strings"
+	"sync/atomic"
+	"time"
+
+	common_errors "github.com/daytonaio/common-go/pkg/errors"
+
+	"github.com/daytonaio/daemon/pkg/childreap"
+	"github.com/daytonaio/daemon/pkg/common"
+	"github.com/gin-gonic/gin"
+)
+
+// ExecuteCommand godoc
+//
+//	@Summary		Execute a command
+//	@Description	Execute a shell command and return the output and exit code
+//	@Tags			process
+//	@Accept			json
+//	@Produce		json
+//	@Param			request	body		ExecuteRequest	true	"Command execution request"
+//	@Success		200		{object}	ExecuteResponse
+//	@Router			/process/execute [post]
+//
+//	@id				ExecuteCommand
+func ExecuteCommand(logger *slog.Logger) gin.HandlerFunc {
+	return func(c *gin.Context) {
+		var request ExecuteRequest
+		if err := c.ShouldBindJSON(&request); err != nil {
+			c.Error(common_errors.NewBadRequestError(fmt.Errorf("invalid request body: %w", err)))
+			return
+		}
+
+		if strings.TrimSpace(request.Command) == "" {
+			c.Error(common_errors.NewBadRequestError(errors.New("command cannot be empty or whitespace-only")))
+			return
+		}
+
+		if err := common.ValidateEnvKeys(request.Envs); err != nil {
+			c.Error(common_errors.NewBadRequestError(err))
+			return
+		}
+
+		cmd := buildExecCmd(request.Command)
+		if request.Cwd != nil {
+			cmd.Dir = *request.Cwd
+		}
+		common.ApplyEnvs(cmd, request.Envs)
+
+		// Capture combined stdout+stderr ourselves so we can route through
+		// childreap.Wait (cmd.CombinedOutput would call cmd.Wait directly
+		// and race the PID-1 reaper).
+		var outBuf bytes.Buffer
+		cmd.Stdout = &outBuf
+		cmd.Stderr = &outBuf
+
+		if err := cmd.Start(); err != nil {
+			c.JSON(http.StatusOK, ExecuteResponse{
+				ExitCode: -1,
+				Result:   err.Error(),
+			})
+			return
+		}
+
+		// set maximum execution time
+		var timeoutReached atomic.Bool
+		if request.Timeout != nil && *request.Timeout > 0 {
+			timeout := time.Duration(*request.Timeout) * time.Second
+			timer := time.AfterFunc(timeout, func() {
+				timeoutReached.Store(true)
+				// Kill the whole process tree so child processes are also terminated
+				if err := killExecProcessGroup(cmd); err != nil {
+					logger.Error("failed to kill process group", "error", err)
+				}
+			})
+			defer timer.Stop()
+		}
+
+		exitCode, waitErr := childreap.Wait(cmd)
+		output := outBuf.Bytes()
+		if timeoutReached.Load() {
+			c.Error(common_errors.NewRequestTimeoutError(errors.New("command execution timeout")))
+			return
+		}
+		if waitErr != nil {
+			c.JSON(http.StatusOK, ExecuteResponse{
+				ExitCode: -1,
+				Result:   string(output),
+			})
+			return
+		}
+
+		c.JSON(http.StatusOK, ExecuteResponse{
+			ExitCode: exitCode,
+			Result:   string(output),
+		})
+	}
+}
